Add support for retires of calls
The CALL_RETRY_COUNT option is useful if you want the library to
compensate for exceptions returned by pylibmc. In case a cache goes
into failed or dead state pylibmc will raise an exception to notify
about the state change. If you set the CALL_RETRY_COUNT to at last 3
the system should be able to compensate for a state transition of

    Working -> Failed -> Dead

without the application noticing it.